requirements.txt
pandas >= 1.5.2
streamlit ~= 1.17
streamlit-aggrid >= 0.3.3
snowflake-connector-python >= 2.9.0
snowflake-sqlalchemy >= 1.4
plotly >= 5.11
